+++ packages/bsb-hdf5/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 7.3.1 (2026-04-16)
+
+### 🩹 Fixes
+
+- Nest installation with Nx ([4bd2458](https://github.com/dbbs-lab/bsb/commit/4bd2458))
+
+### ❤️ Thank You
+
+- drodarie
+
+## 7.3.0 (2026-03-20)
+
+### 🚀 Features
+
+- overwrite labels for ps ([#224](https://github.com/dbbs-lab/bsb/pull/224))
+
+### ❤️ Thank You
+
+- Dimitri RODARIE
+
 ## 7.2.3 (2026-02-21)
 
 This was a version bump only for bsb-hdf5 to align it with other projects, there were no code changes.